% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Returns the sum of two numbers}
\usage{
add(a, b)
}
\arguments{
\item{a}{Number one.}

\item{b}{Number two.}
}
\value{
Sum of two number.
}
\description{
Returns the sum of two numbers
}
\examples{
library(rPackageTutorial)
add(1, 2)
}
