% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{contents}
\alias{contents}
\title{Description of a data frame}
\usage{
contents(data)
}
\arguments{
\item{data}{a data frame.}
}
\value{
a list with 4 components:
}
\description{
\code{contents} provides describes of the contents of a data frame.
}
\examples{
df_info <- contents(happiness)
df_info
plot(df_info)

}
