% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happiness.R
\docType{data}
\name{happiness}
\alias{happiness}
\title{Happiness Datasets}
\format{
An object of class \code{data.frame} with 498 rows and 12 columns.
}
\source{
happiness {zenplots}
}
\usage{
happiness
}
\description{
Happiness Datasets
}
\keyword{datasets}
