% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.contents}
\alias{plot.contents}
\title{Title}
\usage{
\method{plot}{contents}(x, ...)
}
\arguments{
\item{x}{an object of class \code{contents}.}

\item{...}{additional arguments}
}
\value{
a \code{ggplot2} graph
}
\description{
Title
}
\examples{
library(rPackageTutorial)
df_info <- contents(happiness)
plot(df_info)
}
