% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.contents}
\alias{print.contents}
\title{Description of a data frame}
\usage{
\method{print}{contents}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{contents}.}

\item{digits}{number of significant digits to use for percentage.}

\item{...}{additional arguments passed to the print function.}
}
\description{
Description of a data frame
}
\examples{
library(rPackageTutorial)
df_info <- contents(happiness)
print(df_info)
}
